fix: Resolve type conversion errors for iOS compilation
Fixed 10 type errors preventing iOS build:

1. ai_insight.dart:62 - Changed Duration() to DateTime.now()
   for createdAt default value

2. dashboard_screen.dart:193 - Removed const from SizedBox
   containing non-const CustomPaint

3-4. analysis_screen.dart:205-206 - Added .toDouble() for
   position calculations (num → double)

5-6. scoring_screen.dart:238,275 - Added .toInt() for
   arrowCount in List.generate (num → int)

7-8. scoring_screen.dart:292 - Added .toDouble() for arrow
   marker position calculations (num → double)

9-10. session_service.dart:119,125 - Added explicit <int> type
   parameter to fold operations to fix FutureOr<int> type issue

All type conversions ensure proper compilation on iOS platform.

+++ lib/screens/analysis_screen.dart
@@ -202,8 +202,8 @@ class AnalysisScreen extends ConsumerWidget {
               if (hasData)
                 ...stats.heatmapData.take(30).map((position) {
                   return Positioned(
-                    left: 90 + (position.dx * 80),
-                    top: 90 + (position.dy * 80),
+                    left: 90 + (position.dx * 80).toDouble(),
+                    top: 90 + (position.dy * 80).toDouble(),
                     child: Container(
                       width: 8,
                       height: 8,
